backend/requirements.txt
alembic==1.12.1
bcrypt==4.0.1
emails==0.6
email-validator==2.1
fastapi==0.109.1
gunicorn==22.0.0
httpx==0.25.1
jinja2==3.1.4
passlib[bcrypt]==1.7.4
psycopg[binary]==3.1.13
pydantic>2.0
pydantic-settings==2.2.1
pyjwt==2.8.0
python-multipart==0.0.7
sentry-sdk[fastapi]==1.40.6
sqlmodel==0.0.20
tenacity==8.2.3
uvicorn[standard]==0.24

# Development dependencies
coverage==7.4.3
mypy==1.8.0
pre-commit==3.6.2
pytest==7.4.3
ruff==0.2.2
types-passlib==1.7
